Move many of the "core" RNNCells and rnn functions back to TF core.
Unit test files will move in a followup PR.  This is the big API change.
The old behavior (using tf.contrib.rnn....) will continue to work for
backwards compatibility.

PiperOrigin-RevId: 156809677
