+++ tensorflow/contrib/crf/python/ops/crf.py
@@ -41,11 +41,11 @@ from __future__ import print_function
 
 import numpy as np
 
-from tensorflow.contrib.rnn.python.ops import core_rnn_cell
 from tensorflow.python.framework import dtypes
 from tensorflow.python.ops import array_ops
 from tensorflow.python.ops import math_ops
 from tensorflow.python.ops import rnn
+from tensorflow.python.ops import rnn_cell
 from tensorflow.python.ops import variable_scope as vs
 
 __all__ = [
@@ -225,7 +225,7 @@ def crf_binary_score(tag_indices, sequence_lengths, transition_params):
   return binary_scores
 
 
-class CrfForwardRnnCell(core_rnn_cell.RNNCell):
+class CrfForwardRnnCell(rnn_cell.RNNCell):
   """Computes the alpha values in a linear-chain CRF.
 
   See http://www.cs.columbia.edu/~mcollins/fb.pdf for reference.
