+++ tensorflow/contrib/cudnn_rnn/python/kernel_tests/cudnn_rnn_ops_benchmark.py
@@ -22,7 +22,6 @@ import time
 
 from tensorflow.contrib.cudnn_rnn.python.ops import cudnn_rnn_ops
 from tensorflow.contrib.rnn.python.ops import core_rnn
-from tensorflow.contrib.rnn.python.ops import core_rnn_cell_impl
 from tensorflow.contrib.rnn.python.ops import lstm_ops
 from tensorflow.python.client import session
 from tensorflow.python.framework import dtypes
@@ -31,6 +30,7 @@ from tensorflow.python.ops import array_ops
 from tensorflow.python.ops import control_flow_ops
 from tensorflow.python.ops import gradients_impl
 from tensorflow.python.ops import init_ops
+from tensorflow.python.ops import rnn_cell
 from tensorflow.python.ops import variables
 from tensorflow.python.platform import test
 
@@ -131,9 +131,9 @@ class CudnnRNNBenchmark(test.Benchmark):
         ]
         initializer = init_ops.random_uniform_initializer(-0.01, 0.01, seed=127)
 
-        cell = core_rnn_cell_impl.LSTMCell(
+        cell = rnn_cell.LSTMCell(
             num_units=num_units, initializer=initializer, state_is_tuple=True)
-        multi_cell = core_rnn_cell_impl.MultiRNNCell(
+        multi_cell = rnn_cell.MultiRNNCell(
             [cell() for _ in range(num_layers)])
         outputs, final_state = core_rnn.static_rnn(
             multi_cell, inputs, dtype=dtypes.float32)
@@ -159,7 +159,7 @@ class CudnnRNNBenchmark(test.Benchmark):
         ]
         cell = lambda: lstm_ops.LSTMBlockCell(num_units=num_units)  # pylint: disable=cell-var-from-loop
 
-        multi_cell = core_rnn_cell_impl.MultiRNNCell(
+        multi_cell = rnn_cell.MultiRNNCell(
             [cell() for _ in range(num_layers)])
         outputs, final_state = core_rnn.static_rnn(
             multi_cell, inputs, dtype=dtypes.float32)
