+++ tensorflow/contrib/rnn/__init__.py
@@ -16,21 +16,26 @@
 
 See @{$python/contrib.rnn} guide.
 
+# From core
 @@RNNCell
 @@BasicRNNCell
 @@BasicLSTMCell
 @@GRUCell
 @@LSTMCell
-@@LayerNormBasicLSTMCell
 @@LSTMStateTuple
-@@MultiRNNCell
-@@LSTMBlockWrapper
 @@DropoutWrapper
+@@MultiRNNCell
+@@DeviceWrapper
+@@ResidualWrapper
+
+# Used to be in core, but kept in contrib.
 @@EmbeddingWrapper
 @@InputProjectionWrapper
 @@OutputProjectionWrapper
-@@DeviceWrapper
-@@ResidualWrapper
+
+# Created in contrib, eventual plans to move to core.
+@@LayerNormBasicLSTMCell
+@@LSTMBlockWrapper
 @@LSTMBlockCell
 @@GRUBlockCell
 @@FusedRNNCell
@@ -48,9 +53,11 @@ See @{$python/contrib.rnn} guide.
 @@HighwayWrapper
 @@GLSTMCell
 
-### RNNCell wrappers
+# RNNCell wrappers
 @@AttentionCellWrapper
 @@CompiledWrapper
+
+# RNN functions
 @@static_rnn
 @@static_state_saving_rnn
 @@static_bidirectional_rnn
@@ -62,31 +69,23 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from tensorflow.contrib.rnn.python.ops.core_rnn import static_bidirectional_rnn
-from tensorflow.contrib.rnn.python.ops.core_rnn import static_rnn
-from tensorflow.contrib.rnn.python.ops.core_rnn import static_state_saving_rnn
-
-from tensorflow.contrib.rnn.python.ops.core_rnn_cell import BasicLSTMCell
-from tensorflow.contrib.rnn.python.ops.core_rnn_cell import BasicRNNCell
-from tensorflow.contrib.rnn.python.ops.core_rnn_cell import DeviceWrapper
-from tensorflow.contrib.rnn.python.ops.core_rnn_cell import DropoutWrapper
-from tensorflow.contrib.rnn.python.ops.core_rnn_cell import EmbeddingWrapper
-from tensorflow.contrib.rnn.python.ops.core_rnn_cell import GRUCell
-from tensorflow.contrib.rnn.python.ops.core_rnn_cell import InputProjectionWrapper
-from tensorflow.contrib.rnn.python.ops.core_rnn_cell import LSTMCell
-from tensorflow.contrib.rnn.python.ops.core_rnn_cell import LSTMStateTuple
-from tensorflow.contrib.rnn.python.ops.core_rnn_cell import MultiRNNCell
-from tensorflow.contrib.rnn.python.ops.core_rnn_cell import OutputProjectionWrapper
-from tensorflow.contrib.rnn.python.ops.core_rnn_cell import ResidualWrapper
-from tensorflow.contrib.rnn.python.ops.core_rnn_cell import RNNCell
-
 # pylint: disable=unused-import,wildcard-import,line-too-long
+from tensorflow.contrib.rnn.python.ops.core_rnn_cell import EmbeddingWrapper
+from tensorflow.contrib.rnn.python.ops.core_rnn_cell import InputProjectionWrapper
+from tensorflow.contrib.rnn.python.ops.core_rnn_cell import OutputProjectionWrapper
+
 from tensorflow.contrib.rnn.python.ops.fused_rnn_cell import *
 from tensorflow.contrib.rnn.python.ops.gru_ops import *
 from tensorflow.contrib.rnn.python.ops.lstm_ops import *
 from tensorflow.contrib.rnn.python.ops.rnn import *
 from tensorflow.contrib.rnn.python.ops.rnn_cell import *
+
+from tensorflow.python.ops.rnn import static_bidirectional_rnn
+from tensorflow.python.ops.rnn import static_rnn
+from tensorflow.python.ops.rnn import static_state_saving_rnn
+
+from tensorflow.python.ops.rnn_cell import *
 # pylint: enable=unused-import,wildcard-import,line-too-long
 
 from tensorflow.python.util.all_util import remove_undocumented
-remove_undocumented(__name__, ['core_rnn_cell'])
+remove_undocumented(__name__)
