+++ tensorflow/contrib/rnn/python/kernel_tests/lstm_ops_test.py
@@ -20,8 +20,6 @@ from __future__ import print_function
 
 import numpy as np
 
-from tensorflow.contrib.rnn.python.ops import core_rnn
-from tensorflow.contrib.rnn.python.ops import core_rnn_cell_impl
 from tensorflow.contrib.rnn.python.ops import lstm_ops
 from tensorflow.python.framework import constant_op
 from tensorflow.python.framework import dtypes
@@ -30,6 +28,7 @@ from tensorflow.python.ops import array_ops
 from tensorflow.python.ops import gradients_impl
 from tensorflow.python.ops import init_ops
 from tensorflow.python.ops import rnn
+from tensorflow.python.ops import rnn_cell
 from tensorflow.python.ops import variable_scope
 from tensorflow.python.ops import variables
 from tensorflow.python.platform import test
@@ -66,10 +65,9 @@ class LSTMBlockCellTest(test.TestCase):
         m1 = array_ops.zeros([1, 2])
         m2 = array_ops.zeros([1, 2])
         m3 = array_ops.zeros([1, 2])
-        g, ((out_m0, out_m1),
-            (out_m2, out_m3)) = core_rnn_cell_impl.MultiRNNCell(
-                [lstm_ops.LSTMBlockCell(2) for _ in range(2)],
-                state_is_tuple=True)(x, ((m0, m1), (m2, m3)))
+        g, ((out_m0, out_m1), (out_m2, out_m3)) = rnn_cell.MultiRNNCell(
+            [lstm_ops.LSTMBlockCell(2)
+             for _ in range(2)], state_is_tuple=True)(x, ((m0, m1), (m2, m3)))
         sess.run([variables.global_variables_initializer()])
         res = sess.run([g, out_m0, out_m1, out_m2, out_m3], {
             x.name: np.array([[1., 1.]]),
@@ -88,11 +86,11 @@ class LSTMBlockCellTest(test.TestCase):
 
   def testCompatibleNames(self):
     with self.test_session(use_gpu=self._use_gpu, graph=ops.Graph()):
-      cell = core_rnn_cell_impl.LSTMCell(10)
-      pcell = core_rnn_cell_impl.LSTMCell(10, use_peepholes=True)
+      cell = rnn_cell.LSTMCell(10)
+      pcell = rnn_cell.LSTMCell(10, use_peepholes=True)
       inputs = [array_ops.zeros([4, 5])] * 6
-      core_rnn.static_rnn(cell, inputs, dtype=dtypes.float32, scope="basic")
-      core_rnn.static_rnn(pcell, inputs, dtype=dtypes.float32, scope="peephole")
+      rnn.static_rnn(cell, inputs, dtype=dtypes.float32, scope="basic")
+      rnn.static_rnn(pcell, inputs, dtype=dtypes.float32, scope="peephole")
       basic_names = {
           v.name: v.get_shape()
           for v in variables.trainable_variables()
@@ -102,8 +100,8 @@ class LSTMBlockCellTest(test.TestCase):
       cell = lstm_ops.LSTMBlockCell(10)
       pcell = lstm_ops.LSTMBlockCell(10, use_peephole=True)
       inputs = [array_ops.zeros([4, 5])] * 6
-      core_rnn.static_rnn(cell, inputs, dtype=dtypes.float32, scope="basic")
-      core_rnn.static_rnn(pcell, inputs, dtype=dtypes.float32, scope="peephole")
+      rnn.static_rnn(cell, inputs, dtype=dtypes.float32, scope="basic")
+      rnn.static_rnn(pcell, inputs, dtype=dtypes.float32, scope="peephole")
       block_names = {
           v.name: v.get_shape()
           for v in variables.trainable_variables()
@@ -140,11 +138,9 @@ class LSTMBlockCellTest(test.TestCase):
         m1 = array_ops.zeros([1, 2])
         m2 = array_ops.zeros([1, 2])
         m3 = array_ops.zeros([1, 2])
-        g, ((out_m0, out_m1),
-            (out_m2, out_m3)) = core_rnn_cell_impl.MultiRNNCell(
-                [core_rnn_cell_impl.BasicLSTMCell(2, state_is_tuple=True)
-                 for _ in range(2)],
-                state_is_tuple=True)(x, ((m0, m1), (m2, m3)))
+        g, ((out_m0, out_m1), (out_m2, out_m3)) = rnn_cell.MultiRNNCell(
+            [rnn_cell.BasicLSTMCell(2, state_is_tuple=True) for _ in range(2)],
+            state_is_tuple=True)(x, ((m0, m1), (m2, m3)))
         sess.run([variables.global_variables_initializer()])
         basic_res = sess.run([g, out_m0, out_m1, out_m2, out_m3], {
             x.name: x_values,
@@ -159,10 +155,9 @@ class LSTMBlockCellTest(test.TestCase):
         m1 = array_ops.zeros([1, 2])
         m2 = array_ops.zeros([1, 2])
         m3 = array_ops.zeros([1, 2])
-        g, ((out_m0, out_m1),
-            (out_m2, out_m3)) = core_rnn_cell_impl.MultiRNNCell(
-                [lstm_ops.LSTMBlockCell(2) for _ in range(2)],
-                state_is_tuple=True)(x, ((m0, m1), (m2, m3)))
+        g, ((out_m0, out_m1), (out_m2, out_m3)) = rnn_cell.MultiRNNCell(
+            [lstm_ops.LSTMBlockCell(2)
+             for _ in range(2)], state_is_tuple=True)(x, ((m0, m1), (m2, m3)))
         sess.run([variables.global_variables_initializer()])
         block_res = sess.run([g, out_m0, out_m1, out_m2, out_m3], {
             x.name: x_values,
@@ -193,12 +188,12 @@ class LSTMBlockCellTest(test.TestCase):
         m1 = array_ops.zeros([1, 2])
         m2 = array_ops.zeros([1, 2])
         m3 = array_ops.zeros([1, 2])
-        g, ((out_m0, out_m1),
-            (out_m2, out_m3)) = core_rnn_cell_impl.MultiRNNCell(
-                [core_rnn_cell_impl.LSTMCell(
-                    2, use_peepholes=True, state_is_tuple=True)
-                 for _ in range(2)],
-                state_is_tuple=True)(x, ((m0, m1), (m2, m3)))
+        g, ((out_m0, out_m1), (out_m2, out_m3)) = rnn_cell.MultiRNNCell(
+            [
+                rnn_cell.LSTMCell(2, use_peepholes=True, state_is_tuple=True)
+                for _ in range(2)
+            ],
+            state_is_tuple=True)(x, ((m0, m1), (m2, m3)))
         sess.run([variables.global_variables_initializer()])
         basic_res = sess.run([g, out_m0, out_m1, out_m2, out_m3], {
             x.name: x_values,
@@ -213,11 +208,9 @@ class LSTMBlockCellTest(test.TestCase):
         m1 = array_ops.zeros([1, 2])
         m2 = array_ops.zeros([1, 2])
         m3 = array_ops.zeros([1, 2])
-        g, ((out_m0, out_m1),
-            (out_m2, out_m3)) = core_rnn_cell_impl.MultiRNNCell(
-                [lstm_ops.LSTMBlockCell(2, use_peephole=True)
-                 for _ in range(2)],
-                state_is_tuple=True)(x, ((m0, m1), (m2, m3)))
+        g, ((out_m0, out_m1), (out_m2, out_m3)) = rnn_cell.MultiRNNCell(
+            [lstm_ops.LSTMBlockCell(2, use_peephole=True) for _ in range(2)],
+            state_is_tuple=True)(x, ((m0, m1), (m2, m3)))
         sess.run([variables.global_variables_initializer()])
         block_res = sess.run([g, out_m0, out_m1, out_m2, out_m3], {
             x.name: x_values,
@@ -247,8 +240,8 @@ class LSTMBlockCellTest(test.TestCase):
       initializer = init_ops.random_uniform_initializer(
           -0.01, 0.01, seed=19890212)
       with variable_scope.variable_scope("basic", initializer=initializer):
-        cell = core_rnn_cell_impl.BasicLSTMCell(cell_size, state_is_tuple=True)
-        outputs, state = core_rnn.static_rnn(cell, inputs, dtype=dtypes.float32)
+        cell = rnn_cell.BasicLSTMCell(cell_size, state_is_tuple=True)
+        outputs, state = rnn.static_rnn(cell, inputs, dtype=dtypes.float32)
 
         sess.run([variables.global_variables_initializer()])
         basic_outputs, basic_state = sess.run([outputs, state[0]])
@@ -321,9 +314,9 @@ class LSTMBlockCellTest(test.TestCase):
       initializer = init_ops.random_uniform_initializer(
           -0.01, 0.01, seed=19890212)
       with variable_scope.variable_scope("basic", initializer=initializer):
-        cell = core_rnn_cell_impl.LSTMCell(
+        cell = rnn_cell.LSTMCell(
             cell_size, use_peepholes=True, state_is_tuple=True)
-        outputs, state = core_rnn.static_rnn(cell, inputs, dtype=dtypes.float32)
+        outputs, state = rnn.static_rnn(cell, inputs, dtype=dtypes.float32)
 
         sess.run([variables.global_variables_initializer()])
         basic_outputs, basic_state = sess.run([outputs, state[0]])
@@ -410,8 +403,8 @@ class LSTMBlockCellTest(test.TestCase):
       initializer = init_ops.random_uniform_initializer(
           -0.01, 0.01, seed=19890213)
       with variable_scope.variable_scope("basic", initializer=initializer):
-        cell = core_rnn_cell_impl.BasicLSTMCell(cell_size, state_is_tuple=True)
-        outputs, state = core_rnn.static_rnn(
+        cell = rnn_cell.BasicLSTMCell(cell_size, state_is_tuple=True)
+        outputs, state = rnn.static_rnn(
             cell, inputs, dtype=dtypes.float32, sequence_length=seq_lengths)
         sess.run([variables.global_variables_initializer()])
         basic_outputs, basic_state = sess.run([outputs, state[0]])
