+++ tensorflow/contrib/rnn/python/ops/rnn.py
@@ -17,7 +17,6 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from tensorflow.contrib.rnn.python.ops import core_rnn as contrib_rnn
 from tensorflow.python.ops import array_ops
 from tensorflow.python.ops import rnn
 from tensorflow.python.ops import variable_scope as vs
@@ -106,7 +105,7 @@ def stack_bidirectional_rnn(cells_fw,
         initial_state_bw = initial_states_bw[i]
 
       with vs.variable_scope("cell_%d" % i) as cell_scope:
-        prev_layer, state_fw, state_bw = contrib_rnn.static_bidirectional_rnn(
+        prev_layer, state_fw, state_bw = rnn.static_bidirectional_rnn(
             cell_fw,
             cell_bw,
             prev_layer,
